+++ tests/testthat/test-basic.R
@@ -4,7 +4,7 @@ test_that("unjoin works", {
   library(dplyr)
   data("Seatbelts", package= "datasets")
   x <- unjoin(as.data.frame(Seatbelts), front, rear, kms)
-  all.equal(inner_join(x$main, x$data) %>% select(-.idx0), as.data.frame(Seatbelts))
+  all.equal(inner_join(x$.idx0, x$data) %>% select(-.idx0), as.data.frame(Seatbelts))
 
   iris %>% unjoin(-Species)
   chickwts %>% unjoin(weight)
@@ -19,9 +19,12 @@ test_that("unjoin works", {
     unjoin(gapminder)
   }
   x <- unjoin(iris, Petal.Width)
-  x %>%  expect_named(c("main", "data"))
-  x$data %>% inner_join(x$main) %>% expect_named(c("Sepal.Length", "Sepal.Width", "Petal.Length", "Species", ".idx0",
+  x %>%  expect_named(c(".idx0", "data"))
+  x$data %>% inner_join(x$.idx0) %>% expect_named(c("Sepal.Length", "Sepal.Width", "Petal.Length", "Species", ".idx0",
                                                    "Petal.Width"))
+
+  x <- unjoin(iris, Petal.Width, Species, key_col = "vertexes")
+  x %>%  expect_named(c("vertexes", "data"))
 })