+++ NEWS.md
@@ -1,5 +1,7 @@
 # unjoin dev
 
+* first version of class "unjoin" and chaining
+
 * fix bug that was preventing key_col arg from working